Move some test utils into shared header file
